refactor: separate Add/DeleteQuestionButtons
