+++ web/src/components/SurveyForm/ChoiceAnswerForm.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import styled, { DefaultTheme } from 'styled-components';
 
 import { SurveyCreateRequest } from '../../types/request/Survey';
+import EditQuestionButton from './EditQuestionButton';
 import QuestionTypeSelect from './QuestionTypeSelect';
 
 const Container = styled.div``;
@@ -29,32 +30,6 @@ const QuestionDescriptionInput = styled(TextInput).attrs({ type: 'text' })`
   margin-top: 7px;
   margin-bottom: 23px;
   margin-right: 18vw;
-`;
-
-const Button = styled.button`
-  font-weight: 900;
-  text-align: center;
-  padding: 10px;
-  background-color: ${(props) => props.theme.colors.button};
-  width: 35px;
-  height: 35px;
-  border-radius: 50%;
-  border: none;
-  cursor: pointer;
-
-  &:hover {
-    opacity: 0.8;
-  }
-`;
-
-const AddQuestionButton = styled(Button)`
-  color: ${(props) => props.theme.colors.text};
-  margin-left: 37vw;
-`;
-
-const DeleteQuestionButton = styled(Button)`
-  color: #cd5c5c;
-  margin-left: 2vw;
 `;
 
 const AddOptionButton = styled.button`
@@ -103,13 +78,7 @@ export default function ChoiceAnswerForm({
         value={surveyData.questions[questionId].title || ''}
       />
       {QuestionTypeSelect({ selected, questionId, handleChangeQuestionType, theme })}
-      <DeleteQuestionButton
-        theme={theme}
-        name="deleteQuestion"
-        onClick={(event) => handleClickButton(event, questionId)}
-      >
-        X
-      </DeleteQuestionButton>
+      {EditQuestionButton({ editType: 'Delete', questionId, handleClickButton, theme })}
 
       <QuestionDescriptionInput
         theme={theme}
@@ -125,9 +94,7 @@ export default function ChoiceAnswerForm({
       </AddOptionButton>
       <br />
 
-      <AddQuestionButton theme={theme} name="addQuestion" onClick={(event) => handleClickButton(event, questionId)}>
-        +
-      </AddQuestionButton>
+      {EditQuestionButton({ editType: 'Add', questionId, handleClickButton, theme })}
     </Container>
   );
 }
